django_backend/BetterDays/requirements.txt
Pillow
Django
djangorestframework
asgiref==3.8.1
sqlparse==0.5.3
tzdata==2025.1
django-cors-headers==4.2.0
# pillow is needed for images 
gunicorn==20.1.0
